adding some features

+++ src/main/java/com/fatihsengun/service/impl/RefreshTokenServiceImpl.java
@@ -16,6 +16,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -31,7 +32,10 @@ public class RefreshTokenServiceImpl implements IRefreshTokenService {
     @Override
     @Transactional
     public RefreshToken saveRefreshToken(User user) {
-        deleteRefresh(user);
+        Optional<RefreshToken> refreshToken = refreshTokenRepository.findRefreshTokenByUserId(user.getId());
+        if (refreshToken.isPresent()) {
+            deleteRefresh(user);
+        }
 
         RefreshToken newRefreshToken = new RefreshToken();
         newRefreshToken.setUser(user);
